+++ backends/data.go
@@ -21,11 +21,40 @@ type DataInterface interface {
 	// BufferDeviceNum returns the deviceNum for the buffer.
 	BufferDeviceNum(buffer Buffer) DeviceNum
 
-	// BufferToFlatData transfers the flat values of buffer to the Go flat array. The slice flat must have
-	// the exact number of elements required to store the Buffer shape. See BufferShape, and shapes.Shape.Size.
+	// BufferToFlatData transfers the flat values of buffer to the Go flat array.
+	// The slice flat must have the exact number of elements required to store the Buffer shape.
+	//
+	// See also FlatDataToBuffer, BufferShape, and shapes.Shape.Size.
 	BufferToFlatData(buffer Buffer, flat any)
 
 	// BufferFromFlatData transfers data from Go given as a flat slice (of the type corresponding to the shape DType)
 	// to the deviceNum, and returns the corresponding Buffer.
 	BufferFromFlatData(deviceNum DeviceNum, flat any, shape shapes.Shape) Buffer
+
+	// HasSharedBuffers returns whether the backend supports "shared buffers": these are buffers
+	// that can be used directly by the engine and has a local address that can be read or mutated
+	// directly by the client.
+	HasSharedBuffers() bool
+
+	// NewSharedBuffer returns a "shared buffer" that can be both used as input for execution of
+	// computations and directly read or mutated by the clients.
+	//
+	// It panics if the backend doesn't support shared buffers -- see HasSharedBuffer.
+	//
+	// The shared buffer should not be mutated while it is used by an execution.
+	// Also, the shared buffer cannot be "donated" during execution.
+	//
+	// When done, to release the memory, call BufferFinalized on the returned buffer.
+	//
+	// It returns a handle to the buffer and a slice of the corresponding data type pointing
+	// to the shared data.
+	NewSharedBuffer(deviceNum DeviceNum, shape shapes.Shape) (buffer Buffer, flat any)
+
+	// BufferData returns a slice pointing to the buffer storage memory directly.
+	//
+	// This only works if HasSharedBuffer is true, that is, if the backend engine runs on CPU, or
+	// shares CPU memory.
+	//
+	// The returned slice becomes invalid after the buffer is destroyed.
+	BufferData(buffer Buffer) (flat any)
 }